renamed modal -> ExplanationModal
